API

Types

CGcoefficient.HalfInt — Type

HalfInt = Union{Integer, Rational}

Angular momentum quantum numbers may be half integers and integers. With HalfInt type, you can use both 2 and 3//2 as parameters. But the parameter like 5//3, who's denominator is not 2 while gives out error.

CGcoefficient.SqrtRational — Type

SqrtRational <: Real

Store exact value of a Rational's square root. It is stored in s√r format, and we do not simplify it during arithmetics. You can use the simplify function to simplify it.

Core functions

CGcoefficient.CG — Function

CG(j1::HalfInt, j2::HalfInt, j3::HalfInt, m1::HalfInt, m2::HalfInt, m3::HalfInt)

CG coefficient $\langle j_1m_1 j_2m_2 | j_3m_3 \rangle$

CGcoefficient.threeJ — Function

threeJ(j1::HalfInt, j2::HalfInt, j3::HalfInt, m1::HalfInt, m2::HalfInt, m3::HalfInt)

Wigner 3j-symbol

\[\begin{pmatrix} j_1 & j_2 & j_3 \\ m_1 & m_2 & m_3 \end{pmatrix}\]

CGcoefficient.sixJ — Function

sixJ(j1::HalfInt, j2::HalfInt, j3::HalfInt, j4::HalfInt, j5::HalfInt, j6::HalfInt)

Wigner 6j-symbol

\[\begin{Bmatrix} j_1 & j_2 & j_3 \\ j_4 & j_5 & j_6 \end{Bmatrix}\]

CGcoefficient.nineJ — Function

nineJ(j1::HalfInt, j2::HalfInt, j3::HalfInt,
      j4::HalfInt, j5::HalfInt, j6::HalfInt,
      j7::HalfInt, j8::HalfInt, j9::HalfInt)

Wigner 9j-symbol

\[\begin{Bmatrix} j_1 & j_2 & j_3 \\ j_4 & j_5 & j_6 \\ j_7 & j_8 & j_9 \end{Bmatrix}\]

CGcoefficient.Racah — Function

Racah(j1::HalfInt, j2::HalfInt, j3::HalfInt, j4::HalfInt, j5::HalfInt, j6::HalfInt)

Racah coefficient

\[W(j_1j_2j_3j_4, j_5j_6) = (-1)^{j_1+j_2+j_3+j_4} \begin{Bmatrix} j_1 & j_2 & j_5 \\ j_4 & j_3 & j_6 \end{Bmatrix}\]

People often use double of angular momentum quantum number as parameters, so we can use integer as parameters. This package also offers such functions, where the d letter means double.

CGcoefficient.dCG — Function

dCG(dj1::Integer, dj2::Integer, dj3::Integer, dm1::Integer, dm2::Integer, dm3::Integer)

CG coefficient function with double angular monentum number parameters, so that the parameters can be integer. You can calculate dCG(1, 1, 2, 1, 1, 2) to calculate the real CG(1//2, 1//2, 1, 1/2, 1//2, 1)

CGcoefficient.d3j — Function

d3j(dj1::Integer, dj2::Integer, dj3::Integer, dm1::Integer, dm2::Integer, dm3::Integer)

3j-symbol function with double angular monentum number parameters, so that the parameters can be integer.

CGcoefficient.d6j — Function

d6j(dj1::Integer, dj2::Integer, dj3::Integer, dj4::Integer, dj5::Integer, dj6::Integer)

6j-symbol function with double angular monentum number parameters, so that the parameters can be integer.

CGcoefficient.d9j — Function

d9j(dj1::Integer, dj2::Integer, dj3::Integer,
    dj4::Integer, dj5::Integer, dj6::Integer,
    dj7::Integer, dj8::Integer, dj9::Integer)

9j-symbol function with double angular monentum number parameters, so that the parameters can be integer.

CGcoefficient.dRacah — Function

dRacah(dj1::Integer, dj2::Integer, dj3::Integer, dj4::Integer, dj5::Integer, dj6::Integer)

Racah coefficient function with double angular momentum parameters, so that the parameters can be integer.

float version functions

Float version functions is always used for numeric calculation, so the parameters of all these functions (except wigner_init_float) are double of the exact angular momentum quantum number.

CGcoefficient.fbinomial — Function

fbinomial(n::Integer, k::Integer)

binomial with Float64 return value.

CGcoefficient.unsafe_fbinomial — Function

function unsafe_fbinomial(n::Int, k::Int)

Same as fbinomial, but without bounds check. Thus for n < 0 or n > _nmax or k < 0 or k > n, the result is undefined. In Wigner symbol calculation, the mathematics guarantee that unsafe_fbinomial(n, k) is always safe.

CGcoefficient.fCG — Function

fCG(dj1::Integer, dj2::Integer, dj3::Integer, dm1::Integer, dm2::Integer, dm3::Integer)

float64 and fast CG coefficient.

CGcoefficient.f3j — Function

f3j(dj1::Integer, dj2::Integer, dj3::Integer, dm1::Integer, dm2::Integer, dm3::Integer)

float64 and fast Wigner 3j symbol.

CGcoefficient.f6j — Function

f6j(dj1::Integer, dj2::Integer, dj3::Integer, dj4::Integer, dj5::Integer, dj6::Integer)

float64 and fast Wigner 6j symbol.

CGcoefficient.f9j — Function

f9j(dj1::Integer, dj2::Integer, dj3::Integer,
    dj4::Integer, dj5::Integer, dj6::Integer,
    dj7::Integer, dj8::Integer, dj9::Integer)

float64 and fast Wigner 9j symbol.

CGcoefficient.fRacah — Function

Racah(dj1::Integer, dj2::Integer, dj3::Integer, dj4::Integer, dj5::Integer, dj6::Integer)

float64 and fast Racah coefficient.

CGcoefficient.wigner_init_float — Function

wigner_init_float(n::Integer, mode::AbstractString, rank::Integer)

This function reserves memory for fbinomial(n, k). In this code, the fbinomial function is only valid in the stored range. If you call a fbinomial function out of the range, it just gives you 0. The __init__() function stores to nmax = 67.

The parameters means

Calculate range	CG & 3j	6j & Racah	9j

meaning of type	type\rank	3	6	9

max angular momentum	"Jmax"	3*Jmax+1	4*Jmax+1	5*Jmax+1

max two-body coupled angular momentum	"2bjmax"	2*jmax+1	3*jmax+1	4*jmax+1

max binomial	"nmax"	nmax	namx	nmax

The "2bjmax" mode means your calculation only consider two-body coupling, and no three-body coupling. This mode assumes that in all these coefficients, at least one of the angular momentun is just a single particle angular momentum. With this assumption, "2bjmax" mode will use less memory than "Jmax" mode.

"Jmax" means the global maximum angular momentum, for every parameters. It is always safe with out any assumption.

The "nmax" mode directly set nmax, and the rank parameter is ignored.

rank = 6 means you only need to calculate CG and/or 6j symbols, you don't need to calculate 9j symbol.

For example

wigner_init_float(21, "Jmax", 6)

means you calculate CG and 6j symbols, and donot calculate 9j symbol. The maximum angular momentum in your system is 21.

You do not need to rememmber those values in the table. You just need to find the maximum angular momentum in you canculation, then call the function.

The wignerinitfloat function is not thread safe, so you should call it before you start your calculation.

Some useful function

CGcoefficient.iphase — Function

iphase(n::Integer)

$(-1)^n$

CGcoefficient.is_same_parity — Function

is_same_parity(x::T, y::T) where {T <: Integer}

judge if two integers are same odd or same even

CGcoefficient.check_jm — Function

check_jm(dj::T, dm::T) where {T <: Integer}

check if the m-quantum number if one of the components of the j-quantum number, in other words, m and j has the same parity, and abs(m) < j

CGcoefficient.check_couple — Function

check_couple(dj1::T, dj2::T, dj3::T) where {T <: Integer}

check if three angular monentum number j1, j2, j3 can couple

CGcoefficient.exact_sqrt — Function

exact_sqrt(r::Union{Integer, Rational})

Get exact √r using SqrtRational type.

Base.float — Method

float(x::SqrtRational)::Float64

Although we use BigInt in calculation, we do not convert it to BigFloat. Because this package is designed for numeric calculation, BigFloat is unnecessary.

CGcoefficient.simplify — Method

simplify(n::Integer)

Simplify a integer n = x * t^2 to (x, t)

CGcoefficient.simplify — Method

simplify(x::SqrtRational)

Simplify a SqrtRational.